+++ cli/client/filter.go
@@ -4,19 +4,19 @@ import (
 	"encoding/json"
 
 	corev2 "github.com/sensu/sensu-go/api/core/v2"
-	"github.com/sensu/sensu-go/types"
 )
 
-var filtersPath = createNSBasePath(coreAPIGroup, coreAPIVersion, "filters")
+// FiltersPath is the api path for filters.
+var FiltersPath = createNSBasePath(coreAPIGroup, coreAPIVersion, "filters")
 
 // CreateFilter creates a new filter on configured Sensu instance
-func (client *RestClient) CreateFilter(filter *types.EventFilter) (err error) {
+func (client *RestClient) CreateFilter(filter *corev2.EventFilter) (err error) {
 	bytes, err := json.Marshal(filter)
 	if err != nil {
 		return err
 	}
 
-	path := filtersPath(client.config.Namespace())
+	path := FiltersPath(client.config.Namespace())
 	res, err := client.R().SetBody(bytes).Post(path)
 	if err != nil {
 		return err
@@ -31,14 +31,14 @@ func (client *RestClient) CreateFilter(filter *types.EventFilter) (err error) {
 
 // DeleteFilter deletes a filter from configured Sensu instance
 func (client *RestClient) DeleteFilter(namespace, name string) error {
-	return client.Delete(filtersPath(namespace, name))
+	return client.Delete(FiltersPath(namespace, name))
 }
 
 // FetchFilter fetches a specific check
-func (client *RestClient) FetchFilter(name string) (*types.EventFilter, error) {
-	var filter *types.EventFilter
+func (client *RestClient) FetchFilter(name string) (*corev2.EventFilter, error) {
+	var filter *corev2.EventFilter
 
-	path := filtersPath(client.config.Namespace(), name)
+	path := FiltersPath(client.config.Namespace(), name)
 	res, err := client.R().Get(path)
 	if err != nil {
 		return nil, err
@@ -52,25 +52,14 @@ func (client *RestClient) FetchFilter(name string) (*types.EventFilter, error) {
 	return filter, err
 }
 
-// ListFilters fetches all filters from configured Sensu instance
-func (client *RestClient) ListFilters(namespace string, options *ListOptions) ([]corev2.EventFilter, error) {
-	var filters []corev2.EventFilter
-
-	if err := client.List(filtersPath(namespace), &filters, options); err != nil {
-		return filters, err
-	}
-
-	return filters, nil
-}
-
 // UpdateFilter updates an existing filter with fields from a new one.
-func (client *RestClient) UpdateFilter(f *types.EventFilter) error {
+func (client *RestClient) UpdateFilter(f *corev2.EventFilter) error {
 	b, err := json.Marshal(f)
 	if err != nil {
 		return err
 	}
 
-	path := filtersPath(f.Namespace, f.Name)
+	path := FiltersPath(f.Namespace, f.Name)
 	resp, err := client.R().SetBody(b).Put(path)
 	if err != nil {
 		return err